+++ mefit/Utils/FileUtils.cs
@@ -8,6 +8,7 @@ using Mac_EFI_Toolkit.Common;
 using System;
 using System.Diagnostics;
 using System.IO;
+using System.Security.Cryptography;
 
 #region Enum
 enum Status
@@ -21,6 +22,20 @@ namespace Mac_EFI_Toolkit.Utils
 {
     class FileUtils
     {
+        /// <summary>
+        /// Calculates the SHA256 hash of a byte array.
+        /// </summary>
+        /// <param name="sourceBytes">The byte array to calculate the hash for.</param>
+        /// <returns>The SHA256 checksum of the byte array.</returns>
+        internal static string GetSha256Digest(byte[] sourceBytes)
+        {
+            using (SHA256 provider = SHA256.Create())
+            {
+                byte[] digestBytes = provider.ComputeHash(sourceBytes);
+                return BitConverter.ToString(digestBytes).Replace("-", "").ToLower();
+            }
+        }
+
         /// <summary>
         /// Calculates the CRC32 checksum of a byte array. 
         /// </summary>